# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Task",
        "description": "Это API для управления задачами, построенное на Laravel.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить все задачи",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Поиск задач по названию",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Создать новую задачу",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "due_date",
                                    "priority",
                                    "status",
                                    "category"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Изучить Laravel"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Завершить учебник по API Laravel"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-01-20T15:00:00"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "низкий",
                                            "средний",
                                            "высокий"
                                        ],
                                        "example": "высокий"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "в ожидании",
                                            "выполнено"
                                        ],
                                        "example": "в ожидании"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Работа"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Задача успешно создана"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить задачу по ID",
                "operationId": "90b3cd4d2d4b416dc989d2055dbed06c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детали задачи"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Обновить существующую задачу",
                "operationId": "1aaf37e74360ef4c10fadaee02ae77fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Обновленная задача"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "в ожидании",
                                            "выполнено"
                                        ],
                                        "example": "выполнено"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Задача успешно обновлена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удалить задачу",
                "operationId": "ad582480cd57c2a4e3b6a426943eeca5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача успешно удалена"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Всё о задачах"
        }
    ]
}